Add error handling.

+++ FilesHub/test/test.view.jar.files.sh
@@ -1,13 +1,22 @@
 # View all files in fileshub.jar
+FILESHUB_JAR=fileshub.jar
 FILESHUB_TMP_DIR=fileshub
 
-rm -rf ${FILESHUB_TMP_DIR}
-mkdir ${FILESHUB_TMP_DIR}
-cp ${FILESHUB_TMP_DIR}.jar ${FILESHUB_TMP_DIR}
-cd ${FILESHUB_TMP_DIR}
-jar xf ${FILESHUB_TMP_DIR}.jar
-rm -f ${FILESHUB_TMP_DIR}.jar
-cd ..
-clear
-find ${FILESHUB_TMP_DIR} -type f | sort
+if [ ! -f ${FILESHUB_JAR}.jar ]
+then
+  echo "ERROR: ${FILESHUB_JAR}.jar is not found!"
+else
+
+  rm -rf ${FILESHUB_TMP_DIR}
+	mkdir ${FILESHUB_TMP_DIR}
+	cp ${FILESHUB_JAR}.jar ${FILESHUB_TMP_DIR}
+	cd ${FILESHUB_TMP_DIR}
+	jar xf ${FILESHUB_JAR}.jar
+	rm -f ${FILESHUB_JAR}.jar
+	cd ..
+	clear
+	find ${FILESHUB_TMP_DIR} -type f | sort
+            
+fi
+
